requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
selenium==4.15.2
webdriver-manager==4.0.1
slack-sdk==3.27.1
python-dotenv==1.0.0
markdownify==0.11.6
schedule==1.2.0
sqlalchemy==2.0.36
alembic==1.13.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
anthropic 